requirements.txt
PyQt6
adafruit-circuitpython-max31855
pyserial
python-daemon
PyYAML
RPi.GPIO
spidev
